requirements.txt
Django>=4.2,<5.1
spacy>=3.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
numpy
scikit-learn
